# 31-Spring-Boot-REST-API-IRCTC-App/Swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
	"tags": [
		{
			"name": "ticket-rest-controller",
			"description": "Ticket Rest Controller"
		}
	],
	"paths": {
		"/ticket": {
			"post": {
				"tags": [
					"ticket-rest-controller"
				],
				"summary": "bookTicket",
				"operationId": "bookTicketUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "passenger",
						"description": "passenger",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Passenger"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Ticket"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/ticket/{ticketNum}": {
			"get": {
				"tags": [
					"ticket-rest-controller"
				],
				"summary": "getTicket",
				"operationId": "getTicketUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "ticketNum",
						"in": "path",
						"description": "ticketNum",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Ticket"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"Passenger": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"from": {
					"type": "string"
				},
				"jdate": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"phno": {
					"type": "integer",
					"format": "int64"
				},
				"to": {
					"type": "string"
				},
				"trainNum": {
					"type": "integer",
					"format": "int64"
				}
			},
			"title": "Passenger"
		},
		"Ticket": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"from": {
					"type": "string"
				},
				"jdate": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"phno": {
					"type": "integer",
					"format": "int64"
				},
				"ticketNum": {
					"type": "string"
				},
				"ticketPrice": {
					"type": "number",
					"format": "double"
				},
				"ticketStatus": {
					"type": "string"
				},
				"to": {
					"type": "string"
				},
				"trainNum": {
					"type": "integer",
					"format": "int64"
				}
			},
			"title": "Ticket"
		}
	}
}